Additional changes
More changes in order to facilitate simple integration of cibuildwheel for building all the wheels
